added a hostname flag/var and Docker usage docs
Now Wadsworth should be a bit easier to use within a container. This change adds documentation for Docker usage as well as a solution to a problem introduced by using a container: the lack of access to the host's hostname. This change introduces a new flag/environment variable named `hostname`/`HOSTNAME` so the parent machine's hostname can be passed in to wadsworth's container and used during configuration execution.