fixed some major issues with how errors are handled
The old code here ignored initialisation errors. The new code fixes this with some slightly convoluted code. The reason it's a bit complicated is the nature of some errors being immediate and some being sent on a channel. Which must then be wrapped into a second layer of concurrency and multiplexed back into the central for-select loop.