resolve #46
Adds a new option: pass-env, which when true will pass the pico process environment to children. Defaults to false to promote separation of environments.

Adds support for passing prefixed variables from the global Pico. The prefix is GLOBAL_ and is not configurable because I felt the config flags are growing.

Adds some better unit tests for execution config and environment merging.